refactor: ProfileUser, ProfileContents 컴포넌트분리
